data/api/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2
requests
Pillow
cassandra-driver
colorthief
bs4
